Engage kludge/workaround for a very specific ffmpeg warning that is driving me crazy even though it only happens on 0.1% of videos.
If the audio is being converted but the video is not, sometimes ffmpeg will spam warnings about
how there is a "non-monotonous dts in output stream" -- This basically means that the sound is going
to be out of sync with the video and the only way to fix this it to re-encode the video along with the sound.
I don't feel like reorganizing everything to support sending the same file back through ffmpeg with different commands
Instead, we're going to close the current ffmpeg instance, and pipe the file through manual.py with a
new option to force re-encoding.
The script will wait here until the subprocess is finished, in which it then exits this function and
pretends that everything is a-okay so that sabn/nzbget/etc scripts will properly autoimport the file.